Crash scenario with one machine for azure.

+++ lib/containers/bats.pm
@@ -16,7 +16,7 @@ use testapi;
 use utils;
 use strict;
 use warnings;
-use version_utils qw(is_sle);
+use version_utils qw(is_sle is_tumbleweed);
 use serial_terminal qw(select_user_serial_terminal select_serial_terminal);
 use registration qw(add_suseconnect_product get_addon_fullname);
 use bootloader_setup 'add_grub_cmdline_settings';
@@ -24,6 +24,7 @@ use power_action_utils 'power_action';
 use List::MoreUtils qw(uniq);
 use YAML::PP;
 use File::Basename;
+use Utils::Architectures qw(is_aarch64);
 
 our @EXPORT = qw(
   bats_post_hook
@@ -72,22 +73,35 @@ sub install_git {
     # We need git 2.47.0+ to use `--ours` with `git apply -3`
     if (is_sle) {
         my $version = get_var("VERSION");
-        $version =~ s/-SP/./;
-        run_command "sudo zypper addrepo https://download.opensuse.org/repositories/devel:/languages:/python:/backports/$version/devel:languages:python:backports.repo";
+        if (is_sle('<16')) {
+            $version =~ s/-/_/;
+            $version = "SLE_$version";
+        }
+        run_command "sudo zypper addrepo https://download.opensuse.org/repositories/Kernel:/tools/$version/Kernel:tools.repo";
     }
     run_command "sudo zypper --gpg-auto-import-keys -n install --allow-vendor-change git-core", timeout => 300;
 }
 
 sub install_ncat {
-    my $version = get_var("NCAT_VERSION", "7.95-3");
-
-    my @cmds = (
-        "rpm -vhU https://nmap.org/dist/ncat-$version.x86_64.rpm",
-        "ln -sf /usr/bin/ncat /usr/bin/nc"
-    );
-    foreach my $cmd (@cmds) {
-        run_command $cmd;
+    # Upstream tests use Nmap's ncat but this is a problematic tool for multiple reasons:
+    # - Behaviour breaks between versions. See https://nmap.org/changelog.html#7.96
+    # - Nmap changed license and both Fedora & openSUSE won't ship new versions
+    # Tumbleweed has ncat 7.95 and SLES 16.0 has 7.92. These versions are unlikely to change
+    if (is_tumbleweed) {
+        if (is_aarch64) {
+            run_command "zypper addrepo http://download.opensuse.org/ports/aarch64/tumbleweed/repo/non-oss/ non-oss";
+        } else {
+            run_command "zypper addrepo http://download.opensuse.org/tumbleweed/repo/non-oss/ non-oss";
+        }
+    } elsif (is_sle('<16')) {
+        # This repo has ncat 7.92.  Also unlikely to change
+        run_command "zypper addrepo https://download.opensuse.org/repositories/network:/utilities/SLE_15/network:utilities.repo";
     }
+    run_command "zypper --gpg-auto-import-keys -n install ncat";
+
+    # Some tests use nc instead of ncat but expect ncat behaviour instead of netcat-openbsd
+    run_command "ln -sf /usr/bin/ncat /usr/bin/nc";
+    record_info("nc", script_output("nc --version"));
 }
 
 sub install_bats {
@@ -157,7 +171,7 @@ sub enable_modules {
     add_suseconnect_product(get_addon_fullname('desktop'));
     add_suseconnect_product(get_addon_fullname('sdk'));
     add_suseconnect_product(get_addon_fullname('python3')) if is_sle('>=15-SP4');
-    # Needed for criu & fakeroot
+    # Needed for criu, fakeroot & qemu-linux-user
     add_suseconnect_product(get_addon_fullname('phub'));
 }
 
@@ -228,16 +242,22 @@ sub bats_setup {
 
     enable_modules if is_sle;
 
+    my $install_ncat = 0;
+    if (grep { $_ eq "ncat" } @pkgs) {
+        $install_ncat = 1;
+        @pkgs = grep { $_ ne "ncat" } @pkgs;
+    }
+
     # Install tests dependencies
     my $oci_runtime = get_var("OCI_RUNTIME", "");
     if ($oci_runtime && !grep { $_ eq $oci_runtime } @pkgs) {
         push @pkgs, $oci_runtime;
     }
-    run_command "zypper --gpg-auto-import-keys -n install @pkgs";
+    run_command "zypper --gpg-auto-import-keys -n install @pkgs", timeout => 300;
 
     configure_oci_runtime $oci_runtime;
 
-    install_ncat if (get_required_var("BATS_PACKAGE") =~ /^aardvark-dns|netavark|podman$/);
+    install_ncat if $install_ncat;
 
     install_git;
 
@@ -253,8 +273,8 @@ sub bats_setup {
         assert_script_run "semodule -DB || true";
     }
 
-    # Remove mounts.conf
-    run_command "rm -vf /etc/containers/mounts.conf /usr/share/containers/mounts.conf";
+    # This is a workaround for https://bugzilla.suse.com/show_bug.cgi?id=1246227
+    run_command "rm -vf /etc/containers/mounts.conf /usr/share/containers/mounts.conf" unless (is_sle(">=16") || is_tumbleweed);
 
     # Disable tmpfs from next boot
     if (script_output("findmnt -no FSTYPE /tmp", proceed_on_failure => 1) =~ /tmpfs/) {
@@ -276,6 +296,22 @@ sub bats_setup {
     assert_script_run "mount --make-rshared /tmp" if (script_run("findmnt -no FSTYPE /tmp") == 0);
 }
 
+sub collect_coredumps {
+    script_run('coredumpctl list > coredumpctl.txt');
+
+    # Get PID and executable for all dumps
+    my @lines = split /\n/, script_output(q{coredumpctl --no-pager --no-legend | awk '$9 == "present" { print $5, $10 }'});
+
+    foreach my $line (@lines) {
+        my ($pid, $exe) = split /\s+/, $line;
+        my $core = basename($exe) . ".$pid.core";
+
+        # Dumping and compressing coredumps may take some time
+        script_run("coredumpctl -o $core dump $pid", 300);
+        script_run("gzip -v $core", 300);
+    }
+}
+
 sub bats_post_hook {
     select_serial_terminal;
 
@@ -285,6 +321,7 @@ sub bats_post_hook {
 
     script_run "rm -rf $test_dir";
 
+    collect_coredumps;
     script_run('df -h > df-h.txt');
     script_run('dmesg > dmesg.txt');
     script_run('findmnt > findmnt.txt');
@@ -359,7 +396,7 @@ sub bats_tests {
 
     unless (get_var("BATS_TESTS")) {
         $skip_tests = get_var($skip_tests, $settings->{$skip_tests});
-        my @skip_tests = split(/\s+/, get_var('BATS_SKIP', $settings->{BATS_SKIP}) . " " . $skip_tests);
+        my @skip_tests = split(/\s+/, join(' ', get_var('BATS_SKIP', $settings->{BATS_SKIP}), $skip_tests));
         patch_logfile($log_file, @skip_tests);
     }
 
@@ -368,36 +405,6 @@ sub bats_tests {
     run_command "sudo rm -rf $tmp_dir || true";
 
     return ($ret);
-}
-
-sub bats_patches {
-    return if check_var("BATS_PATCHES", "none");
-
-    my $github_org = ($package eq "runc") ? "opencontainers" : "containers";
-
-    my @patches = split(/\s+/, get_var("BATS_PATCHES", ""));
-    if (!@patches && defined $settings->{BATS_PATCHES}) {
-        @patches = @{$settings->{BATS_PATCHES}};
-    }
-
-    foreach my $patch (@patches) {
-        my $url = ($patch =~ /^\d+$/) ? "https://github.com/$github_org/$package/pull/$patch.patch" : $patch;
-        record_info("patch", $url);
-        if ($patch =~ /^\d+$/) {
-            push @commands, "curl $curl_opts -O $url";
-            assert_script_run "curl -O " . data_url("containers/bats/patches/$package/$patch.patch");
-        } else {
-            run_command "curl $curl_opts -O $url", timeout => 900;
-        }
-        # Some patches (e.g., podman's 25942) fail to apply cleanly due to missing files so
-        # try `git apply` first and if it fails use `--include` to restrict the patch scope
-        # to the package tests directory.  Remove this when `git-apply` has a new option to
-        # ignore missing files.
-        my $file = basename($url);
-        my $apply_cmd = "git apply -3 --ours $file";
-        $apply_cmd .= " || git apply -3 --ours --include '$tests_dir{$package}/*' $file";
-        run_command $apply_cmd;
-    }
 }
 
 sub bats_settings {
@@ -435,7 +442,32 @@ sub bats_sources {
     run_command "cd $test_dir";
     run_command "git checkout $branch";
 
-    bats_patches;
+    # We use BATS_PATCHES="none" to specify that we don't want to patch anything
+    unless (check_var("BATS_PATCHES", "none")) {
+        my @patches = split(/\s+/, get_var("BATS_PATCHES", ""));
+        if (!@patches && defined $settings->{BATS_PATCHES}) {
+            @patches = @{$settings->{BATS_PATCHES}};
+        }
+
+        foreach my $patch (@patches) {
+            my $url = ($patch =~ /^\d+$/) ? "https://github.com/$github_org/$package/pull/$patch.patch" : $patch;
+            record_info("patch", $url);
+            if ($patch =~ /^\d+$/) {
+                push @commands, "curl $curl_opts -O $url";
+                assert_script_run "curl -O " . data_url("containers/bats/patches/$package/$patch.patch");
+            } else {
+                run_command "curl $curl_opts -O $url", timeout => 900;
+            }
+            # Some patches (e.g., podman's 25942) fail to apply cleanly due to missing files so
+            # try `git apply` first and if it fails use `--include` to restrict the patch scope
+            # to the package tests directory.  Remove this when `git-apply` has a new option to
+            # ignore missing files.
+            my $file = basename($url);
+            my $apply_cmd = "git apply -3 --ours $file";
+            $apply_cmd .= " || git apply -3 --ours --include '$tests_dir{$package}/*' $file";
+            run_command $apply_cmd;
+        }
+    }
 
     if ($package eq "podman") {
         my $hack_bats = "https://raw.githubusercontent.com/containers/podman/refs/heads/main/hack/bats";
